[refactor] move build_sla_func_tensor to tzrec.ops.attention_utils
The helper builds the NFUNC=3 mask tensor consumed by both the CUTLASS
production path (cutlass_hstu_mha) and the PyTorch reference path
(pytorch_hstu_mha via _decode_attn_func_to_mask); it has no CUTLASS-
specific dependencies. Move it to a backend-agnostic location so the
name reflects scope and neither backend module imports CUTLASS-only
code just to get the helper.

New module: ``tzrec.ops.attention_utils`` with a short preamble
explaining the intent.

Import updates: STULayer.forward, STUStack.forward, and three sites in
hstu_attention_test.py now import from ``tzrec.ops.attention_utils``.

Also fold HstuSLAAttnTest / PytorchAttnFuncTest / HstuAttnNegativeTests
back into HSTUAttentionTest so every HSTU attention test lives under
one class. The top-level ``test_sla_attn`` parity helper stays where
it is (module-level, not a method).
